Own testing module for decompositions
